Add optional `embedded-hal-async` support
This commit adds support for the `embedded-hal-async` crate in addition
to `embedded-hal`. I've done this by adding a separate `AsyncSgp30`
type, based on the assumption that most projects won't need to use both
the blocking `embedded-hal` traits and the `embedded-hal-async` traits
at the same time, and providing `async fn` methods on a separate type
with the same names as the blocking ones seemed a bit nicer than having
one type that has async methods suffixed with `_async` or similar.

In order to make this change, I've factored out some code into functions
on `MeasurmentData` and `RawConfig` so that they can be shared between
the blocking and async implementations.

Support for `embedded-hal-async` is gated behind the
`embedded-hal-async` feature flag, so the dependency is not enabled by
default. Therefore, I've also added documentation about the feature
flag, and configured docs.rs to build the documentation with the
`doc_cfg` and `doc_auto_cfg` features enabled, so that the documentation
clearly shows which types require the feature flag.

Thanks for implementing the driver! Please let me know if you have any
questions about this PR.